Reorganized code for better best practices
This moves the static content into the "static" folder and moves
the javascript code into a .js file.

This also optionally allows the user to specify the API_KEY in an
environment variable. Will check that this will be an easy place
to specify / point to in the IDX template setup.

The client and model creation are moved to the main function so they can be reused.

The images are collected as a Glob pattern and added some sanity
checking on the image names to prevent users being able to ask it
questions about arbitrary files on the server.

The server should now be run from the root folder:
  go run ./cmd/web